Added Identify Plane, BidsViewer and RunCount. Version 0.1.0

+++ Batch_AddStudy.py
@@ -8,10 +8,66 @@ from concurrent.futures import ThreadPoolExecutor
 import multiprocessing
 import pandas as pd
 from datetime import datetime, timedelta
+import argparse
 import glob
 
+"""
+DICOM Processing Script
+
+This script processes DICOM files for medical imaging research. It supports anonymization,
+sorting, conversion to BIDS format, and populating participant data. Various steps in the process can be skipped using 
+command line flags.
+
+Usage:
+    python script_name.py [options]
+
+Options:
+    --noanon          Skip the anonymization step.
+    --nosort          Skip the sorting step.
+    --nobids          Skip the conversion to BIDS format.
+    --nocleanup       Skip cleanup of temporary unsorted anonymized dicom dir.
+    --noparticipants  Skip populating participants.tsv file.
+"""
+
+
+
+def parse_arguments():
+    parser = argparse.ArgumentParser(
+        description="Processes DICOM files for medical imaging research.",
+        formatter_class=argparse.RawTextHelpFormatter
+    )
+    parser.add_argument(
+        '--noanon', 
+        action='store_true', 
+        help='Skip the anonymization step (use if data is already anonymized).'
+    )
+    parser.add_argument(
+        '--nosort', 
+        action='store_true', 
+        help='Skip the sorting step (use if sorting is not required).'
+    )
+    parser.add_argument(
+        '--nobids', 
+        action='store_true', 
+        help='Skip the conversion to BIDS format (use if conversion is already done or not needed).'
+    )
+
+    parser.add_argument(
+        '--nocleanup', 
+        action='store_true', 
+        help='Skip cleanup of temporary unsorted anonymized dicom dir.'
+    )
+
+    parser.add_argument(
+        '--noparticipants',
+        action='store_true',
+        help='Skip populating participants.tsv file.'
+    )
+    return parser.parse_args()
+
+
 def read_subject_mapping(filename):
-    df = pd.read_csv(filename, sep='\s+', header=None, dtype=str)
+    df = pd.read_csv(filename, sep=r'\s+', header=None, dtype=str)
     # Map DICOM PatientID to new subject ID (e.g., '41411412222222': 'sub-001')
     return {row[1]: 'sub-' + row[0].zfill(3) for _, row in df.iterrows()}
 
@@ -109,22 +165,67 @@ def process_new_sessions(sourcedata_dir, bidsdir_folder, dcm2bids_config):
 
 
 
+def populate_participants_tsv(inbox_folder, participant_map_file, bidsdir_folder):
+    participant_map = read_subject_mapping(participant_map_file)
+    participants_file = os.path.join(bidsdir_folder, "participants.tsv")
+
+    required_columns = ["participant_id", "age", "sex", "notes"]
+    should_append = file_has_required_columns(participants_file, required_columns)
+
+    if not should_append:
+        with open(participants_file, "w") as file:
+            file.write("participant_id\tage\tsex\tgroup\tnotes\toriginal_id\n")
+
+    participants_data = {}  # Initialize the participants_data dictionary
+
+    with open(participants_file, "a" if should_append else "w") as file:
+        if not should_append:
+            file.write("participant_id\tage\tsex\tgroup\tnotes\toriginal_id\n")
+
+        for root, dirs, files in os.walk(inbox_folder):
+            for file_name in files:
+                if is_dicom_file(file_name):
+                    dicom_file = os.path.join(root, file_name)
+                    ds = pydicom.dcmread(dicom_file)
+                    original_patient_id = ds.PatientID if 'PatientID' in ds else "Unknown"
+                    age = ds.PatientAge if 'PatientAge' in ds else ""
+                    sex = ds.PatientSex if 'PatientSex' in ds else ""
+                    new_patient_id = participant_map.get(original_patient_id, "Unknown")
+                    if new_patient_id not in participants_data:
+                        participants_data[new_patient_id] = [new_patient_id, age, sex, "", "", original_patient_id]
+                    break  # Break after processing the first DICOM file in each folder
+
+        for data in participants_data.values():
+            file.write("\t".join(data) + "\n")
+
+def file_has_required_columns(file_path, required_columns):
+    """Check if a file has the required columns."""
+    try:
+        df = pd.read_csv(file_path, sep='\t')
+        return all(column in df.columns for column in required_columns)
+    except Exception:
+        return False
+        
 
 def main():
+    args = parse_arguments()
     bidsfolder = os.path.dirname(os.path.realpath(__file__))
     os.chdir(bidsfolder)
 
-    subject_map = read_subject_mapping("ID_correspondence.tsv")
+    
 
+    # Read subject mapping
+    subject_map = read_subject_mapping("ID_correspondence.tsv")
     patient_id_map = subject_map
 
-
+    # Directory setup
     bidsdir_folder = os.path.join(bidsfolder, "BIDSDIR")
     sourcedata_dir = os.path.join(bidsdir_folder, "sourcedata")
     raw_dicom_folder = os.path.join(bidsfolder, "Inbox")
     anon_dicom_folder = os.path.join(bidsfolder, ".temp_anondir")
     dcm2bids_config = "dcm2bids_config.json"
 
+    # BIDS directory setup
     if not os.path.exists(bidsdir_folder):
         print("Creating BIDS directory structure.")
         scaffold_cmd = ["dcm2bids_scaffold", "-o", bidsdir_folder]
@@ -135,17 +236,37 @@ def main():
     os.makedirs(sourcedata_dir, exist_ok=True)
     os.makedirs(anon_dicom_folder, exist_ok=True)
 
-    process_directory(raw_dicom_folder, anon_dicom_folder, patient_id_map)
 
 
-    print("Sorting DICOM files.")
-    dicom_sort_cmd = ["dicom_sort", "--copy", anon_dicom_folder, sourcedata_dir, "%PatientID%/%StudyDate%/%SeriesDescription%"]
-    subprocess.run(dicom_sort_cmd)
+    # Run participants_data.py script
+    if not args.noparticipants:
+        print("Populating participants.tsv file.")
+        populate_participants_tsv(os.path.join(bidsfolder, "Inbox"), 
+                                  "ID_correspondence.tsv", 
+                                  os.path.join(bidsfolder, "BIDSDIR"))
 
-    process_new_sessions(sourcedata_dir, bidsdir_folder, dcm2bids_config)
- 
-    shutil.rmtree(anon_dicom_folder)
+
+
+    # Anonymization step
+    if not args.noanon:
+        process_directory(raw_dicom_folder, anon_dicom_folder, patient_id_map)
+    
+    # Sorting step
+    if not args.nosort:
+        print("Sorting DICOM files.")
+        dicom_sort_cmd = ["dicom_sort", "--copy", anon_dicom_folder, sourcedata_dir, "%PatientID%/%StudyDate%/%SeriesDescription%"]
+        subprocess.run(dicom_sort_cmd)
+
+    # dcm2bids step
+    if not args.nobids:
+        process_new_sessions(sourcedata_dir, bidsdir_folder, dcm2bids_config)
+
+    if not args.nocleanup:
+        shutil.rmtree(anon_dicom_folder)
+
     print("Batch process completed.")
+
+
 
 if __name__ == "__main__":
     main()
